Reorganize into modules.
