+++ src/main.rs
@@ -1,21 +1,10 @@
-mod ast;
-mod flat_ast;
-mod generator;
-mod lexer;
-mod tokens;
+mod s0_lexer;
+mod s1_parser;
+mod s2_analyzer;
+mod s3_optimizer;
+mod s4_generator;
 
 use clap::Parser;
-use generator::flatten;
-use generator::parse_file;
-
-use lalrpop_util::lalrpop_mod;
-
-lalrpop_mod!(
-    #[allow(clippy::ptr_arg)]
-    #[allow(clippy::vec_box)]
-    #[rustfmt::skip]
-    parser
-);
 
 #[derive(Parser, Debug)]
 #[command(version, about = "Rumoca Modelica Translator", long_about = None)]
@@ -35,13 +24,13 @@ struct Args {
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let args = Args::parse();
-    let def = parse_file(&args.filename).expect("failed to parse");
-    let flat_def = flatten(&def).expect("failed to flatten");
+    let def = s1_parser::parse_file(&args.filename).expect("failed to parse");
+    let flat_def = s2_analyzer::flatten(&def).expect("failed to flatten");
 
     if args.verbose {
         println!("{:#?}", flat_def);
     }
-    let s = generator::generate(&flat_def, &args.template_file)?;
+    let s = s4_generator::generate(&flat_def, &args.template_file)?;
     println!("{s:}");
     Ok(())
 }
