+++ src/s0_lexer/lexer.rs
@@ -1,6 +1,6 @@
 use logos::{Logos, SpannedIter};
 
-use crate::tokens::{LexicalError, Token}; // your Token enum, as above
+use crate::s0_lexer::tokens::{LexicalError, Token}; // your Token enum, as above
 
 pub type Spanned<Tok, Loc, Error> = Result<(Loc, Tok, Loc), Error>;
