+++ src/s2_analyzer/mod.rs
@@ -1,0 +1,4 @@
+pub mod ast;
+pub mod flattener;
+
+pub use flattener::flatten;
