+++ src/s4_generator/mod.rs
@@ -1,0 +1,2 @@
+pub mod generator;
+pub use generator::generate;
